рефакторинг вьюшек
