+++ src/main.js
@@ -1,5 +1,4 @@
 import TripPresenter from './presenter/trip-presenter.js';
-import 'flatpickr/dist/flatpickr.min.css';
 
 const tripControlsContainer = document.querySelector('.trip-controls__filters');
 const tripEventsContainer = document.querySelector('.trip-events');
